refactor: share workspace file matcher between code analysis and LS
