+++ crates/emmylua_ls/src/context/workspace_manager.rs
@@ -14,8 +14,8 @@ use crate::context::ServerContextSnapshot;
 use crate::context::lsp_features::LspFeatures;
 use crate::handlers::{ClientConfig, init_analysis, register_files_watch};
 use emmylua_code_analysis::{
-    EmmyLuaAnalysis, Emmyrc, WorkspaceFolder, build_workspace_folders, find_library_exclude,
-    load_configs, read_file_with_encoding, update_code_style, uri_to_file_path,
+    EmmyLuaAnalysis, Emmyrc, WorkspaceFileMatcher, WorkspaceFolder, load_configs,
+    read_file_with_encoding, update_code_style, uri_to_file_path,
 };
 use lsp_types::Uri;
 use tokio::sync::{Mutex as AsyncMutex, RwLock};
@@ -592,206 +592,6 @@ impl WorkspaceDiagnosticLevel {
 
     pub fn to_u8(self) -> u8 {
         self as u8
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct WorkspaceFileMatcher {
-    include: Vec<String>,
-    entries: Vec<WorkspaceMatchEntry>,
-    watch_roots: HashSet<PathBuf>,
-}
-
-#[derive(Debug, Clone)]
-struct WorkspaceMatchEntry {
-    root: PathBuf,
-    is_library: bool,
-    exclude: Vec<String>,
-    exclude_dir: Vec<PathBuf>,
-}
-
-impl WorkspaceFileMatcher {
-    pub fn new(workspace_folders: &[WorkspaceFolder], emmyrc: &Emmyrc) -> Self {
-        let (include, exclude, exclude_dir) =
-            emmylua_code_analysis::calculate_include_and_exclude(emmyrc);
-        let mut entries = build_workspace_folders(workspace_folders, emmyrc)
-            .into_iter()
-            .flat_map(|workspace| {
-                WorkspaceMatchEntry::from_workspace(workspace, &exclude, &exclude_dir, emmyrc)
-            })
-            .collect::<Vec<_>>();
-
-        add_child_workspace_excludes(&mut entries);
-        let watch_roots = entries.iter().map(|entry| entry.root.clone()).collect();
-        entries.sort_by_key(|entry| {
-            std::cmp::Reverse((entry.root.components().count(), entry.is_library))
-        });
-
-        Self {
-            include,
-            entries,
-            watch_roots,
-        }
-    }
-
-    pub fn is_match(&self, path: &std::path::Path) -> bool {
-        let include_set = match wax::any(self.include.iter().map(|s| s.as_str())) {
-            Ok(include_set) => include_set,
-            Err(_) => {
-                log::error!("Invalid include pattern");
-                return true;
-            }
-        };
-
-        for entry in &self.entries {
-            let Ok(relative_path) = path.strip_prefix(&entry.root) else {
-                continue;
-            };
-
-            if entry.exclude_dir.iter().any(|dir| path.starts_with(dir)) {
-                continue;
-            }
-
-            if !entry.exclude.is_empty() {
-                match wax::any(entry.exclude.iter().map(|s| s.as_str())) {
-                    Ok(exclude_set) if wax::Pattern::is_match(&exclude_set, relative_path) => {
-                        continue;
-                    }
-                    Ok(_) => {}
-                    Err(_) => log::error!("Invalid exclude pattern"),
-                }
-            }
-
-            if wax::Pattern::is_match(&include_set, relative_path) {
-                return true;
-            }
-        }
-
-        false
-    }
-
-    pub fn watch_roots(&self) -> HashSet<PathBuf> {
-        self.watch_roots.clone()
-    }
-
-    pub fn source_file_globs(&self) -> &[String] {
-        &self.include
-    }
-}
-
-impl WorkspaceMatchEntry {
-    fn from_workspace(
-        workspace: WorkspaceFolder,
-        exclude: &[String],
-        exclude_dir: &[PathBuf],
-        emmyrc: &Emmyrc,
-    ) -> Vec<Self> {
-        let is_library = workspace.is_library;
-        let override_parent_ignore_dir = workspace.override_parent_ignore_dir;
-        let mut exclude = exclude.to_vec();
-        let mut exclude_dir = if workspace.overrides_parent_ignore_dir() {
-            exclude_dir
-                .iter()
-                .filter(|dir| match &workspace.import {
-                    emmylua_code_analysis::WorkspaceImport::All => !workspace.root.starts_with(dir),
-                    emmylua_code_analysis::WorkspaceImport::SubPaths(paths) => !paths
-                        .iter()
-                        .map(|path| workspace.root.join(path))
-                        .any(|path| path.starts_with(dir)),
-                })
-                .cloned()
-                .collect::<Vec<_>>()
-        } else {
-            exclude_dir.to_vec()
-        };
-        if is_library {
-            let (library_exclude, library_exclude_dir) = find_library_exclude(&workspace, emmyrc);
-            exclude.extend(library_exclude);
-            exclude.sort();
-            exclude.dedup();
-
-            exclude_dir.extend(library_exclude_dir);
-            exclude_dir.sort();
-            exclude_dir.dedup();
-        }
-        let roots = match workspace.import {
-            emmylua_code_analysis::WorkspaceImport::All => vec![workspace.root],
-            emmylua_code_analysis::WorkspaceImport::SubPaths(paths) => paths
-                .into_iter()
-                .map(|path| workspace.root.join(path))
-                .collect(),
-        };
-
-        roots
-            .into_iter()
-            .map(|root| {
-                Self::new(
-                    root,
-                    is_library,
-                    override_parent_ignore_dir,
-                    &exclude,
-                    &exclude_dir,
-                )
-            })
-            .collect()
-    }
-
-    fn new(
-        root: PathBuf,
-        is_library: bool,
-        override_parent_ignore_dir: bool,
-        exclude: &[String],
-        exclude_dir: &[PathBuf],
-    ) -> Self {
-        let exclude_dir = if is_library || override_parent_ignore_dir {
-            exclude_dir
-                .iter()
-                .filter(|dir| !root.starts_with(dir))
-                .cloned()
-                .collect()
-        } else {
-            exclude_dir.to_vec()
-        };
-
-        Self {
-            root,
-            is_library,
-            exclude: exclude.to_vec(),
-            exclude_dir,
-        }
-    }
-}
-
-fn add_child_workspace_excludes(entries: &mut [WorkspaceMatchEntry]) {
-    let roots = entries
-        .iter()
-        .map(|entry| entry.root.clone())
-        .collect::<Vec<_>>();
-    for (idx, entry) in entries.iter_mut().enumerate() {
-        for (other_idx, root) in roots.iter().enumerate() {
-            if idx == other_idx {
-                continue;
-            }
-
-            if let Ok(relative) = root.strip_prefix(&entry.root)
-                && relative.components().count() > 0
-            {
-                entry.exclude_dir.push(root.clone());
-            }
-        }
-
-        entry.exclude_dir.sort();
-        entry.exclude_dir.dedup();
-    }
-}
-
-impl Default for WorkspaceFileMatcher {
-    fn default() -> Self {
-        Self {
-            include: vec!["**/*.lua".to_string()],
-            entries: Vec::new(),
-            watch_roots: HashSet::new(),
-        }
     }
 }
